+++ tests/data/testvar.cpp
@@ -59,7 +59,7 @@ int main(int argc, char *argv[])
   /* Create the output name (for storage of dump files) */
 
   VectorString subparts = separateKeywords(argv[1]);
-  int nargs = subparts.size();
+  int nargs = (int) subparts.size();
   String outname = concatenateStrings("", subparts[nargs-2], subparts[nargs-1], "-");
   ASerializable::setContainerName(true);
   ASerializable::setPrefixName(outname);
